+++ app/Helpers/CollectionHelpers.php
@@ -2,6 +2,8 @@
 
 namespace App\Helpers;
 
+use Illuminate\Support\Collection;
+
 class CollectionHelpers
 {
     /**
@@ -11,6 +13,35 @@ class CollectionHelpers
      */
     public static function collectApi(mixed $value = null)
     {
-        return new \App\Libraries\Api\Models\ApiCollection($value);
+        return new class($value) extends Collection
+        {
+            protected ?Collection $metadata = null;
+
+            public function getMetadata(mixed $name = null): mixed
+            {
+                if (! $this->metadata instanceof Collection) {
+                    return null;
+                }
+
+                if ($name) {
+                    return $this->metadata->get($name);
+                }
+
+                return $this->metadata;
+            }
+
+            public function setMetadata(array|Collection $data): static
+            {
+                $data = $data instanceof Collection
+                    ? $data
+                    : collect($data);
+
+                $this->metadata = $this->metadata instanceof Collection
+                    ? $this->metadata->merge($data)
+                    : $data;
+
+                return $this;
+            }
+        };
     }
 }
