--- app/Repositories/Api/ArtworkRepository.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App\Repositories\Api;
-
-use App\Models\Api\Artwork;
-
-class ArtworkRepository extends BaseApiRepository
-{
-    public function __construct(Artwork $model)
-    {
-        $this->model = $model;
-    }
-}
